moved logic to hooks
